requirement.txt
# ---- CORE ----
numpy==1.26.4
scipy==1.11.4

# ---- MACHINE LEARNING ----
tensorflow==2.16.1
keras==3.2.0
scikit-learn==1.3.2

# ---- IMAGE PROCESSING ----
opencv-python==4.8.1.78
pillow==10.2.0

# ---- OCR / LANGUAGE DETECTION ----
pytesseract==0.3.10

# ---- WEB BACKEND ----
Flask==3.0.2
flask-cors==4.0.0

# ---- MODEL SAVING ----
joblib==1.3.2

# ---- OPTIONAL / SAFE ----
matplotlib==3.7.1
protobuf==4.25.3
h5py==3.10.0


For backend API routes.

✔ flask-cors

To allow frontend (5500) to call backend (5000).

✔ tensorflow + keras

To load feature_extractor.keras and mobilenet_model.h5.

✔ scikit-learn + joblib

For loading svm_model.pkl and knn_model.pkl.

✔ Pillow

To open handwriting images.

✔ numpy

To preprocess images.

✔ opencv-python

Useful for image operations (if needed for future upgrades).

✔ matplotlib

Only needed if you want server-side chart generation.
